new-rag-test/nxgn/requirements.txt
google-generativeai
fitz
pymupdf
chromadb
langchain
sentence-transformers
langchain_openai
langchain_chroma
langchain_community
langchain_huggingface
sentence-transformers
